requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary  # For PostgreSQL
openai           # Or the library for your chosen LLM (e.g., transformers)
streamlit
python-dotenv